% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_log.R
\name{activity_log}
\alias{activity_log}
\title{Retrieve Activity Log Entries}
\usage{
activity_log(
  client_id,
  secret,
  start = NULL,
  end = NULL,
  limit = NULL,
  offset = NULL,
  user_id = NULL
)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{start}{The start time(milliseconds) of when you want to receive log events. \strong{\emph{Optional}}}

\item{end}{The end time(milliseconds) of when you want to receive log events. \strong{\emph{Optional}}}

\item{limit}{The maximum number of events you want to retrieve(default is 50, maximum of 1000). \strong{\emph{Optional}}}

\item{offset}{The offset location of events you retrieve(default is 0). \strong{\emph{Optional}}}

\item{user_id}{Specify the ID of the user to retrieve a certain individual's activities. \strong{\emph{Optional}}}
}
\description{
This function retrieves activity log entries and supports custom limit and offset values. All time/date values are in UTC.
\href{https://developer.domo.com/portal/i19jain6fvwjj-activity-log-api}{DOMO Documentation}.
}
\details{
Optional query parameters may be left as \code{NULL}. If provided, query parameters must be single, non-\code{NA} values.
}
\examples{
activity_log(client_id = client_id,
  secret = secret,
  start = 1709337594000,
  end = 1709937594000,
  limit = 50,
  offset = 0)
}
