% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appdb_collection_delete.R
\name{appdb_collection_delete}
\alias{appdb_collection_delete}
\title{Delete AppDB Collection}
\usage{
appdb_collection_delete(
  collection_id,
  developer_token,
  instance
)
}
\arguments{
\item{collection_id}{The AppDB collection ID.}

\item{developer_token}{A valid Domo Developer Token.}

\item{instance}{Domo instance URL. For example:
\code{"https://company.domo.com"}. A bare hostname is also
accepted; \code{https://} is added automatically when no scheme
is present.}
}
\value{
The API response object returned by
\code{httr2::req_perform()}.
}
\description{
Deletes an existing AppDB collection, along with all documents it
contains. This is destructive and cannot be reversed.
}
\examples{
response <- appdb_collection_delete(
  collection_id = "12345678-1234-1234-1234-123456789012",
  developer_token = developer_token,
  instance = "https://company.domo.com"
)
}
