% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appdb_datastore_create.R
\name{appdb_datastore_create}
\alias{appdb_datastore_create}
\title{Create AppDB Datastore}
\usage{
appdb_datastore_create(name, developer_token, instance)
}
\arguments{
\item{name}{Name of the datastore to create.}

\item{developer_token}{A valid Domo Developer Token.}

\item{instance}{Domo instance URL. For example:
\code{"https://company.domo.com"}. A bare hostname is also
accepted; \code{https://} is added automatically when no scheme
is present.}
}
\value{
The API response object returned by
  \code{httr2::req_perform()}. The created datastore's ID is
  available at \code{httr2::resp_body_json(response)$id}.
}
\description{
Creates a new AppDB datastore, owned by the calling developer
token's user. A datastore is the top-level container that AppDB
collections live in; a brand-new collection not tied to an
existing DDX app needs a \code{datastore_id} from here (or from
\code{appdb_datastore_get_all()}) before it can be created.
}
\examples{
response <- appdb_datastore_create(
  name = "A New Datastore",
  developer_token = developer_token,
  instance = "https://company.domo.com"
)

}
