% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appdb_doc_create.R
\name{appdb_doc_create}
\alias{appdb_doc_create}
\title{Create AppDB Document}
\usage{
appdb_doc_create(
  url,
  developer_token,
  data
)
}
\arguments{
\item{url}{The AppDB collection URL.}

\item{developer_token}{A valid Domo Developer Token.}

\item{data}{A document payload, typically generated using
\code{appdb_row_to_document()}.}
}
\description{
Creates a new AppDB document using the provided payload.
}
\value{
The API response object returned by
\code{httr2::req_perform()}.
}
\examples{
new_data <- appdb_row_to_document(
  data = clients,
  row_number = 1
)

response <- appdb_doc_create(
  url = collection_url,
  developer_token = developer_token,
  data = new_data
)
}
