% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appdb_doc_delete.R
\name{appdb_doc_delete}
\alias{appdb_doc_delete}
\title{Delete AppDB Document}
\usage{
appdb_doc_delete(
  collection_id,
  appdb_document_id,
  developer_token,
  instance
)
}
\arguments{
\item{collection_id}{The AppDB collection ID.}

\item{appdb_document_id}{The document ID to delete.}

\item{developer_token}{A valid Domo Developer Token.}

\item{instance}{Domo instance URL. For example:
\code{"https://company.domo.com"}.}
}
\description{
Deletes an existing AppDB document.
}
\value{
The API response object returned by
\code{httr2::req_perform()}.
}
\examples{
response <- appdb_doc_delete(
  collection_id = "12345678-1234-1234-1234-123456789012",
  appdb_document_id = "12345678-1234-1234-1234-123456789012",
  developer_token = developer_token,
  instance = "https://company.domo.com"
)
}
