% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_create.R
\name{dataset_create}
\alias{dataset_create}
\title{Create An Empty DataSet}
\usage{
dataset_create(client_id, secret, name, description, schema)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{name}{Name of the DataSet to create.}

\item{description}{Description of DataSet to create.}

\item{schema}{The current schema associated with this DataSet. A tibble/dataframe that has two columns: "name" & "type". The "name" field is the list of column names and "type" is the type of each column such as:
\itemize{
  \item STRING
  \item DECIMAL
  \item LONG
  \item DOUBLE
  \item DATE
  \item DATETIME}}
}
\description{
Creates a new empty DataSet in your Domo instance. Once the empty DataSet has been created, data can then be imported into the DataSet. Use \code{pRoDsCreate()} function for the best option.
\href{https://developer.domo.com/portal/0d0517e6b6830-create-a-data-set}{DOMO Documentation}.
}
\examples{
dataset_create(client_id = client_id,
  secret = secret,
  name = 'A New Dataset',
  description = 'Description of the new dataset',
  schema = data.frame(name =c('col1', 'col2'), type = c('STRING', 'DOUBLE')))
}
