% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_delete.R
\name{dataset_delete}
\alias{dataset_delete}
\title{Delete a DataSet}
\usage{
dataset_delete(client_id, secret, dataset_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{dataset_id}{The unique ID of the dataset.}
}
\description{
\strong{This is destructive and cannot be reversed.} Permanently deletes a DataSet from your Domo instance. This can be done for all DataSets, not just those created through the API.
\href{https://developer.domo.com/portal/8e96d3ae1061d-delete-a-data-set}{DOMO Documentation}.
}
\examples{
dataset_delete(client_id = client_id,
  secret = secret,
  dataset_id = '3c907f1e-846c-4d06-88f6-35592db151f4')
}
