% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_export.R
\name{dataset_export}
\alias{dataset_export}
\title{Export data from a DOMO DataSet}
\usage{
dataset_export(client_id, secret, dataset_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{dataset_id}{The unique ID of the dataset.}
}
\description{
Export data from a DataSet in your Domo instance. Data types will be exported as they are currently stored in the dataset. In addition, the only supported export type is CSV.
\href{https://developer.domo.com/portal/83ef8a4d71450-export-data-from-data-set}{DOMO Documentation}.
}
\examples{
dataset_export(client_id = client_id,
  secret = secret,
  dataset_id = '3c907f1e-846c-4d06-88f6-35592db151f4')
}
