% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_get_info.R
\name{dataset_get_info}
\alias{dataset_get_info}
\title{Retrieve a DataSet}
\usage{
dataset_get_info(client_id, secret, dataset_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{dataset_id}{The unique ID of the dataset.}
}
\description{
Retrieves the details of an existing DataSet.
\href{https://developer.domo.com/portal/ea210df52a8a1-retrieve-a-data-set}{DOMO Documentation}.
}
\examples{
user_update(client_id = client_id,
  secret = secret,
  dataset_id = '3c907f1e-846c-4d06-88f6-35592db151f4')
}
