% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_get_info_all.R
\name{dataset_get_info_all}
\alias{dataset_get_info_all}
\title{List DataSets}
\usage{
dataset_get_info_all(client_id, secret, nameLike = NULL, sort = NULL)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{nameLike}{If included, will limit the list of DataSets to those with names that contain the string passed in. nameLike is case insensitive. \strong{\emph{Optional}}}

\item{sort}{The DataSet field to sort by. Fields prefixed with a negative sign reverses the sort (i.e. '-name' does a reverse sort by the name of the DataSets). \strong{\emph{Optional}}}
}
\description{
Get a list of all DataSets in your Domo instance.
\href{https://developer.domo.com/portal/72ae9b3e80374-list-data-sets}{DOMO Documentation}.
}
\examples{
dataset_get_info_all(client_id = client_id,
secret = secret
nameLike = 'First',
sort = 'name')
}
