% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_import.R
\name{dataset_import}
\alias{dataset_import}
\title{Import data into DataSet}
\usage{
dataset_import(client_id, secret, dataset_id, data_table)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can created on the \url{developer.domo.com} page.}

\item{dataset_id}{The unique ID of the dataset.}

\item{data_table}{A dataframe/tibble. The function will convert the data_table into a csv before uploading.}
}
\description{
Import data into a DataSet in your Domo instance. This request will replace the data currently in the DataSet.
\href{https://developer.domo.com/portal/4ba1c85b30ae1-import-data-into-data-set}{DOMO Documentation}.
}
\examples{
dataset_import(client_id = client_id,
  secret = secret,
  dataset_id = '3c907f1e-846c-4d06-88f6-35592db151f4',
  data_table = data_table)
}
