% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_create.R
\name{group_create}
\alias{group_create}
\title{Create a group}
\usage{
group_create(client_id, secret, name)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{name}{The name of the group.}
}
\description{
Creates a new group in your Domo instance.
\href{https://developer.domo.com/portal/d88967f3c2bd5-create-a-group}{DOMO Documentation}.
}
\examples{
group_create(client_id = client_id,
  secret = secret,
  name = 'Group Name')
}
