% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_delete.R
\name{group_delete}
\alias{group_delete}
\title{Delete a group}
\usage{
group_delete(client_id, secret, group_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{group_id}{The unique ID of the group.}
}
\description{
\strong{This is destructive and cannot be reversed.} Permanently deletes a group from your Domo instance.
\href{https://developer.domo.com/portal/b268da0278c85-delete-a-group}{DOMO Documentation}.
}
\examples{
group_delete(client_id = client_id,
  secret = secret,
  group_id = 12345)
}
