% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_get.R
\name{group_get}
\alias{group_get}
\title{Retrieve a group}
\usage{
group_get(client_id, secret, group_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{group_id}{The unique ID of the group.}
}
\description{
Retrieves the details of an existing group.
\href{https://developer.domo.com/portal/1f17fb4f14577-retrieve-a-group}{DOMO Documentation}.
}
\examples{
group_get(client_id = client_id,
  secret = secret,
  group_id = '123456789')
}
