% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_get_all.R
\name{group_get_all}
\alias{group_get_all}
\title{List groups}
\usage{
group_get_all(client_id, secret)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}
}
\description{
Get a list of all groups in your Domo instance.
\href{https://developer.domo.com/portal/d7364df8a5f06-list-groups}{DOMO Documentation}.
}
\examples{
group_get_all(client_id = client_id, secret = secret)
}
