% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_user_add.R
\name{group_user_add}
\alias{group_user_add}
\title{Add a user to a group}
\usage{
group_user_add(client_id, secret, group_id, user_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{group_id}{The unique Id of the group to add the user to.}

\item{user_id}{The unique Id of the user to be added to the group.}
}
\description{
Add user to a group in your Domo instance.
\href{https://developer.domo.com/portal/95de1a11dfe97-add-a-user-to-a-group}{DOMO Documentation}.
}
\examples{
group_user_add(client_id = client_id,
  secret = secret,
  group_id = '768473',
  user_id = '123456789')
}
