% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_user_get.R
\name{group_user_get}
\alias{group_user_get}
\title{List users in a group}
\usage{
group_user_get(client_id, secret, group_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{group_id}{The unique ID of the group.}
}
\description{
List the users in a group in your Domo instance.
\href{https://developer.domo.com/portal/1ce9b0022073a-list-users-in-a-group}{DOMO Documentation}.
}
\examples{
group_user_get(client_id = client_id,
  secret = secret,
  group_id = '123456789')
}
