% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_user_remove.R
\name{group_user_remove}
\alias{group_user_remove}
\title{Remove a user from a group}
\usage{
group_user_remove(client_id, secret, group_id, user_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{group_id}{The unique Id of the group to add the user to.}

\item{user_id}{The unique Id of the user to be added to the group.}
}
\description{
Remove a user from a group in your Domo instance.
\href{https://developer.domo.com/portal/608c69894df01-remove-a-user-from-a-group}{DOMO Documentation}.
}
\examples{
group_user_remove(client_id = client_id,
  secret = secret,
  group_id = '768473',
  user_id = '123456789')
}
