% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_collection_create.R
\name{page_collection_create}
\alias{page_collection_create}
\title{Create a page collection}
\usage{
page_collection_create(client_id, secret, page_id, body)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{page_id}{The ID of the page.}

\item{body}{An R \code{list()} that contains the specified definitions of the new collection such as:
\itemize{
\item title: Page collection's name displayed above the set of cards.
\item description: Additional text within the page collection. \strong{\emph{Optional}}.
\item cardIds: IDs provided will add or remove cards that are not a part of a page collection.}}
}
\description{
Create a page collection.
\href{https://developer.domo.com/portal/f95f2f9dc8c0f-create-a-page-collection}{DOMO Documentation}.
}
\examples{
cardIds <- list(); cardIds[[1]] <- 96460956
body <- list(title = 'Collection1', description = 'Just a description', cardIds = cardIds)
page_collection_create(client_id = client_id,
  secret = secret,
  page_id = 1805162873,
  body = body)
}
