% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_collection_update.R
\name{page_collection_update}
\alias{page_collection_update}
\title{Update a page collection}
\usage{
page_collection_update(client_id, secret, page_id, page_collection_id, body)
}
\arguments{
\item{client_id}{A client_id that can be created on the developer.domo.com page.}

\item{secret}{A secret that can be created on the developer.domo.com page.}

\item{page_id}{The ID of the page.}

\item{page_collection_id}{The ID of the page collection.}

\item{body}{The parameters of the page collection.}
}
\description{
Update a page collection.
}
\examples{
cardIds <- list(); cardIds[[1]] <- 96460956
body <- list(title = 'Collection1', description = 'Just a description', cardIds = cardIds)
page_collection_update(client_id = client_id,
  secret = secret,
  page_id = 1805162873,
  page_collection_id = 522185994,
  body = body)
}
