% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_create.R
\name{page_create}
\alias{page_create}
\title{Create a page}
\usage{
page_create(client_id, secret, body)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{body}{An R \code{list()} that contains the specified definitions of the new page such as:
\itemize{
  \item \strong{name:} The name of the page.
  \item \strong{parentId:} If provided, the page will be created as a subpage to the page provided.
  \item \strong{locked:} will restrict other users the ability to make edits to page or its content - the default value is false.
  \item \strong{cardIds:} IDs provided will add or remove cards that are not a part of a page collection.
  \item \strong{visibility:} Determines the access given to both individual users or groups within Domo by providing these 2 parameters:
  \itemize{
    \item \strong{userIds:} The IDs of users that will be given access to view the page.
    \item \strong{roupIds:} The IDs of groups that will be given access to view the page.}}}
}
\description{
Creates a new page in your Domo instance.
\href{https://developer.domo.com/portal/0fce9616dfe5c-create-a-page}{DOMO Documentation}.
}
\examples{
page_create(client_id = client_id,
  secret = secret,
  body = list(name = 'Page #1'))
}
