% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_delete.R
\name{page_delete}
\alias{page_delete}
\title{Delete a page}
\usage{
page_delete(client_id, secret, page_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{page_id}{The ID of the page.}
}
\description{
\strong{This is destructive and cannot be reversed.} Permanently deletes a page from your Domo instance.
\href{https://developer.domo.com/portal/d7c166334953f-delete-a-page}{DOMO Documentation}.
}
\examples{
page_delete(client_id = client_id,
  secret = secret,
  page_id = 123455667)
}
