% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_get.R
\name{page_get}
\alias{page_get}
\title{Retrieve a page}
\usage{
page_get(client_id, secret, page_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{page_id}{The ID of the page.}
}
\description{
Retrieves the details of an existing page.
\href{https://developer.domo.com/portal/5fd5b9d881e6b-retrieve-a-page}{DOMO Documentation}.
}
\examples{
page_get(client_id = client_id,
  secret = secret,
  page_id = 123345556)
}
