% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_update.R
\name{page_update}
\alias{page_update}
\title{Update a page}
\usage{
page_update(client_id, secret, page_id, body)
}
\arguments{
\item{client_id}{A client_id that can be created on the developer.domo.com page.}

\item{secret}{A secret that can created on the developer.domo.com page.}

\item{name}{The name of the group.}
}
\description{
Updates the specified page by providing values to parameters passed. Any parameter left out of the request will cause the specific page’s attribute to remain unchanged.
Also, collections cannot be added or removed via this endpoint, only reordered. Giving access to a user or group will also cause that user or group to have access to the parent page (if the page is a subpage). Moving a page by updating the parentId will also cause everyone with access to the page to have access to the new parent page.
}
\examples{
page_update(client_id = client_id,
  secret = secret,
  user_id = '123456789',
  body = list(name = 'New Name'))
}
