% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp_create.R
\name{pdp_create}
\alias{pdp_create}
\title{Create a Personalized Data Permission (PDP) Policy}
\usage{
pdp_create(client_id, secret, dataset_id, body)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{dataset_id}{The unique ID of the dataset.}

\item{body}{Parameters to update the policy.}
}
\description{
Create a PDP policy for user and or group access to data within a DataSet. Users and groups must exist before creating PDP policy. The number of characters for the list of values for a single PDP policy, including a delimiter character for each value, must be less than 255 characters.
\href{https://developer.domo.com/portal/57eb007324b49-create-a-personalized-data-permission-pdp-policy}{DOMO Documentation}.
}
\examples{
pdp_create(client_id = client_id,
  secret = secret,
  dataset_id = '3c907f1e-846c-4d06-88f6-35592db151f4',
  body = body)
}
