% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp_delete.R
\name{pdp_delete}
\alias{pdp_delete}
\title{Delete a Personalized Data Permission (PDP) Policy}
\usage{
pdp_delete(client_id, secret, dataset_id, pdp_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can created on the \url{developer.domo.com} page.}

\item{dataset_id}{The unique ID of the dataset.}

\item{pdp_id}{The unique ID of the policy}
}
\description{
\strong{This is destructive and cannot be reversed.} Permanently deletes a PDP policy on a DataSet in your Domo instance.
\href{https://developer.domo.com/portal/1d5e64ae6ce1a-delete-a-personlized-data-permission-pdp-policy}{DOMO Documentation}.
}
\examples{
pdp_delete(client_id = client_id,
  secret = secret,
  dataset_id = '3c907f1e-846c-4d06-88f6-35592db151f4',
  pdp_id = 123)
}
