% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp_get.R
\name{pdp_get}
\alias{pdp_get}
\title{Retrieve a Personalized Data Permission (PDP) policy}
\usage{
pdp_get(client_id, secret, dataset_id, pdp_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{dataset_id}{The unique ID of the dataset.}

\item{pdp_id}{The unique ID of the policy}
}
\description{
Retrieve a policy from a DataSet within Domo. A DataSet is required for a PDP policy to exist.
\href{https://developer.domo.com/portal/b342ce1054c3a-retrieve-a-personalized-data-permission-pdp-policy}{DOMO Documentation}.
}
\examples{
pdp_get(client_id = client_id,
  secret = secret,
  dataset_id = '3c907f1e-846c-4d06-88f6-35592db151f4',
  pdp_id = 123)
}
