% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp_get_all.R
\name{pdp_get_all}
\alias{pdp_get_all}
\title{List Personalized Data Permission (PDP) policies}
\usage{
pdp_get_all(client_id, secret, dataset_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{dataset_id}{The unique ID of the dataset.}
}
\description{
Retrieve a policy from a DataSet within Domo. A DataSet is required for a PDP policy to exist.
\href{https://developer.domo.com/portal/4cab29aca0fd3-list-personalized-data-permission-pdp-policies}{DOMO Documentation}.
}
\examples{
pdp_get(client_id = client_id,
  secret = secret,
  dataset_id = '3c907f1e-846c-4d06-88f6-35592db151f4')
}
