% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp_update.R
\name{pdp_update}
\alias{pdp_update}
\title{Update a Personalized Data Permission (PDP) policy}
\usage{
pdp_update(client_id, secret, dataset_id, pdp_id, body)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can created on the \url{developer.domo.com} page.}

\item{dataset_id}{The unique ID of the dataset.}

\item{pdp_id}{The unique ID of the policy.}

\item{body}{Parameters to update the policy.}
}
\description{
Update the specific PDP policy for a DataSet by providing values to parameters passed.
\href{https://developer.domo.com/portal/dcdf6cd230d61-update-a-personalized-data-permission-pdp-policy}{DOMO Documentation}.
}
\examples{
pdp_update(client_id = client_id,
  secret = secret,
  dataset_id = '3c907f1e-846c-4d06-88f6-35592db151f4',
  pdp_id = 123,
  body = body)
}
