% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_attachment_download.R
\name{plt_attachment_download}
\alias{plt_attachment_download}
\title{Download attachment}
\usage{
plt_attachment_download(
  client_id,
  secret,
  project_id,
  list_id,
  task_id,
  attachment_id
)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can created on the \url{developer.domo.com} page.}

\item{project_id}{The unique ID of the project.}

\item{list_id}{The unique ID of the list.}

\item{task_id}{The unique ID of the task.}

\item{attachment_id}{The unique ID of the attachment.}
}
\description{
Downloads an individual attachment given an attachment id.
\href{https://developer.domo.com/portal/6b0b830a02e85-download-attachment}{DOMO Documentation}.
}
\examples{
plt_attachment_download(client_id = client_id,
  secret = secret,
  project_id = 123,
  list_id = 5,
  task_id = 12,
  attachment_id = 1)
}
