% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_attachment_get.R
\name{plt_attachment_get}
\alias{plt_attachment_get}
\title{Retrieve a list of attachments}
\usage{
plt_attachment_get(client_id, secret, project_id, list_id, task_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can created on the \url{developer.domo.com} page.}

\item{project_id}{The unique ID of the project.}

\item{list_id}{The unique ID of the list.}

\item{task_id}{The unique ID of the task.}
}
\description{
Retrieve details about all of the attachments belonging to a particular task.
\href{https://developer.domo.com/portal/a571924a55d11-retrieve-list-of-attachments}{DOMO Documentation}.
}
\examples{
plt_attachment_get(client_id = client_id,
  secret = secret,
  project_id = 123,
  list_id = 5,
  task_id = 12)
}
