% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_list_delete.R
\name{plt_list_delete}
\alias{plt_list_delete}
\title{Delete a list}
\usage{
plt_list_delete(client_id, secret, project_id, list_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{project_id}{The unique ID of the project.}

\item{list_id}{The unique ID of the list.}
}
\description{
\strong{This is destructive and cannot be reversed.} Permanently deletes a list from your Domo instance.
\href{https://developer.domo.com/portal/38324ee605641-delete-a-list}{DOMO Documentation}.
}
\examples{
plt_list_delete(client_id = client_id,
  secret = secret,
  project_id = 123,
  list_id = 1)
}
