% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_list_get.R
\name{plt_list_get}
\alias{plt_list_get}
\title{Retrieve Individual List}
\usage{
plt_list_get(client_id, secret, project_id, list_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can created on the \url{developer.domo.com} page.}

\item{project_id}{The unique ID of the project.}

\item{list_id}{The unique ID of the list.}
}
\description{
Retrieves the details of an individual list given a project id and a list id.
\href{https://developer.domo.com/portal/456b2b958eb45-retrieve-individual-list}{DOMO Documentation}.
}
\examples{
plt_list_get(client_id = client_id,
  secret = secret,
  project_id = 123,
  list_id = 1)
}
