% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_list_get_all.R
\name{plt_list_get_all}
\alias{plt_list_get_all}
\title{Retrieve all project lists}
\usage{
plt_list_get_all(client_id, secret, project_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{project_id}{The unique ID of the project.}
}
\description{
Retrieves all lists available within a given project id.
\href{https://developer.domo.com/portal/38e89edd950ec-retrieve-all-project-lists}{DOMO Documentation}.
}
\examples{
plt_list_get_all(client_id = client_id,
  secret = secret,
  project_id = 123)
}
