% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_list_update.R
\name{plt_list_update}
\alias{plt_list_update}
\title{Update a list}
\usage{
plt_list_update(client_id, secret, project_id, list_id, body)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{project_id}{The unique ID of the project.}

\item{list_id}{The unique ID of the list.}

\item{body}{The content of the list. Please refer to the documentation for details.}
}
\description{
Update the details of a list given an existing project id and list id.
\href{https://developer.domo.com/portal/2f759304cdacc-update-a-list}{DOMO Documentation}.
}
\examples{
plt_list_update(client_id = client_id,
  secret = secret,
  project_id = 123,
  list_id = 5,
  body = body)
}
