% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_project_create.R
\name{plt_project_create}
\alias{plt_project_create}
\title{Create a project}
\usage{
plt_project_create(client_id, secret, body)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{body}{The parameters needed to create the project.}
}
\description{
Create a new project in your Domo instance.
\href{https://developer.domo.com/portal/abaea94e18a2f-create-a-project}{DOMO Documentation}.
}
\examples{
members = list(); members[[1]] <- 12345
body <- list(name = 'Project Name', members = members, public = TRUE)
plt_project_create(client_id = client_id,
  secret = secret,
  body = body)
}
