% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_project_delete.R
\name{plt_project_delete}
\alias{plt_project_delete}
\title{Delete a project}
\usage{
plt_project_delete(client_id, secret, project_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{project_id}{The unique ID of the project.}
}
\description{
\strong{This is destructive and cannot be reversed.} Permanently deletes a project from your Domo instance.
\href{https://developer.domo.com/portal/f884f3d008a49-delete-a-project}{DOMO Documentation}.
}
\examples{
plt_project_delete(client_id = client_id,
  secret = secret,
  project_id = 123)
}
