% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_project_get.R
\name{plt_project_get}
\alias{plt_project_get}
\title{Retrieve individual project}
\usage{
plt_project_get(client_id, secret, project_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{project_id}{The unique ID of the project.}
}
\description{
Retrieves the details of an individual existing project given a project id. Use the special project ID "project_id = me" to return your personal project.
\href{https://developer.domo.com/portal/975aea6d35782-retrieve-individual-project}{DOMO Documentation}.
}
\examples{
plt_project_get(client_id = client_id,
  secret = secret,
  project_id = 123)
}
