% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_project_get_all.R
\name{plt_project_get_all}
\alias{plt_project_get_all}
\title{Retrieve all projects}
\usage{
plt_project_get_all(client_id, secret)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}
}
\description{
Retrieves a list of all projects that the client scope has access to.
\href{https://developer.domo.com/portal/631f21c3c5f5f-retrieve-all-projects}{DOMO Documentation}.
}
\examples{
plt_project_get_all(client_id = client_id, secret = secret)
}
