% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_project_member_get.R
\name{plt_project_member_get}
\alias{plt_project_member_get}
\title{Retrieve project members}
\usage{
plt_project_member_get(client_id, secret, project_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{project_id}{The unique ID of the project.}
}
\description{
Retrieves a list of ids of the users that are members of the given project id.
\href{https://developer.domo.com/portal/193cb98fe288b-retrieve-project-members}{DOMO Documentation}.
}
\examples{
plt_project_member_get(client_id = client_id,
  secret = secret,
  project_id = 123)
}
