% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_project_member_update.R
\name{plt_project_member_update}
\alias{plt_project_member_update}
\title{Update project members}
\usage{
plt_project_member_update(client_id, secret, project_id, body)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can created on the \url{developer.domo.com} page.}

\item{project_id}{The unique ID of the project.}

\item{body}{A list of the user IDs.}
}
\description{
Update the members of a given project id.
\href{https://developer.domo.com/portal/3259632a2c091-update-project-members}{DOMO Documentation}.
}
\examples{
members <- list(); members[[1]] <- 790204705
plt_project_member_update(client_id = client_id,
  secret = secret,
  project_id = 123,
  body = members)
}
