% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_project_update.R
\name{plt_project_update}
\alias{plt_project_update}
\title{Retrieve individual project}
\usage{
plt_project_update(client_id, secret, project_id, body)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can created on the \url{developer.domo.com} page.}

\item{project_id}{The unique ID of the project.}

\item{body}{The parameters needed to create the project.}
}
\description{
Updates attributes of an existing project in your Domo instance. The following properties are read-only and cannot be updated with this request: id, members, createdBy, & createdDate.
\href{https://developer.domo.com/portal/bc780d297a0d6-update-a-project}{DOMO Documentation}.
}
\examples{
body <- list(name = 'Project Name', public = TRUE)
plt_project_update(client_id = client_id,
  secret = secret,
  project_id = 123,
  body = body)
}
