% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_task_create.R
\name{plt_task_create}
\alias{plt_task_create}
\title{Create a task}
\usage{
plt_task_create(client_id, secret, project_id, list_id, body)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{project_id}{The unique ID of the project.}

\item{list_id}{The unique ID of the list.}

\item{body}{The content of the task. Please refer to the documentation for details.}
}
\description{
Add a task to a project list.
\href{https://developer.domo.com/portal/e5d77a7e640ac-create-a-task}{DOMO Documentation}.
}
\examples{
contributors <- list(); contributors[[1]] <- 790204705
body <- list(projectId = 7, projectListId = 22, taskName = 'New Task 1', createdDate = paste0(Sys.Date(),'T00:00:00Z'),
             updatedDate = paste0(Sys.Date(),'T00:00:00Z'), priority = 1, createdBy = 790204705, ownedBy = 790204705,
             contributors = contributors, attachmentCount = 0, tags = list(), archived = FALSE)
plt_task_create(client_id = client_id,
  secret = secret,
  project_id = 7,
  list_id = 22,
  body = body)
}
