% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_task_get.R
\name{plt_task_get}
\alias{plt_task_get}
\title{Retrieve individual task}
\usage{
plt_task_get(client_id, secret, project_id, list_id, task_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{project_id}{The unique ID of the project.}

\item{list_id}{The unique ID of the list.}

\item{task_id}{The unique ID of the task.}
}
\description{
Retrieves an individual task from a given project id and list id.
\href{https://developer.domo.com/portal/7e6100b8c7ac9-retrieve-individual-task}{DOMO Documentation}.
}
\examples{
plt_task_get(client_id = client_id,
  secret = secret,
  project_id = 123,
  list_id = 1,
  task_id = 4)
}
