% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_task_get_all.R
\name{plt_task_get_all}
\alias{plt_task_get_all}
\title{Retrieve all project tasks}
\usage{
plt_task_get_all(client_id, secret, project_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{project_id}{The unique ID of the project.}
}
\description{
Retrieves all tasks from a given project id.
\href{https://developer.domo.com/portal/cf1cad1dc8fe5-retrieve-all-project-tasks}{DOMO Documentation}.
}
\examples{
plt_task_get_all(client_id = client_id,
  secret = secret,
  project_id = 123)
}
