% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_task_update.R
\name{plt_task_update}
\alias{plt_task_update}
\title{Update a Task}
\usage{
plt_task_update(client_id, secret, project_id, list_id, task_id, body)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{project_id}{The unique ID of the project.}

\item{list_id}{The unique ID of the list.}

\item{task_id}{The unique ID of the task.}

\item{body}{The content of the task. Please refer to the documentation for details.}
}
\description{
Update the details of a task given an existing project id, list id, and task id.
\href{https://developer.domo.com/portal/a9c700565ffba-update-a-task}{DOMO Documentation}.
}
\examples{
plt_task_update(client_id = client_id,
  secret = secret,
  project_id = 123,
  list_id = 1,
  task_id = 4,
  body = body)
}
