% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRoDsCreate.R
\name{pRoDsCreate}
\alias{pRoDsCreate}
\title{Create A New Dataset Via Stream API}
\usage{
pRoDsCreate(
  client_id,
  secret,
  dataset,
  dataset_name,
  dataset_description = "",
  update_method = "REPLACE",
  parallel = FALSE,
  n_core = NULL
)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{dataset}{The data frame that will be uploaded to DOMO.}

\item{dataset_name}{The name of the dataset.}

\item{dataset_description}{(Optional) The description of the dataset.}

\item{update_method}{The method in which the dataset will be updated, either "REPLACE" or "APPEND".}

\item{parallel}{(Optional) If TRUE, the function will use multiple cores to upload the data frame. `n_core` is required when `parallel` = TRUE.}

\item{n_core}{(Optional) The number of cores to use from your device. It's recommended to leave one core free. If you have 8 cores, enter `n_core` = 7.}
}
\description{
It is best practice to use this function to create a new dataset as it leverages STREAM API that loads data parts at scale.
}
\examples{
pRoDsCreate(client_id = client_id,
  secret = secret,
  dataset = dataset,
  dataset_name = 'The Name of the Dataset',
  dataset_description = 'This is the description',
  update_method = 'REPLACE')
}
