% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRoDsGet.R
\name{pRoDsGet}
\alias{pRoDsGet}
\title{Query a DataSet Through Parallelism}
\usage{
pRoDsGet(
  client_id,
  secret,
  dataset_id,
  sql_query = NULL,
  parallel = FALSE,
  n_core = NULL
)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can created on the \url{developer.domo.com} page.}

\item{dataset_id}{The unique ID of the dataset.}

\item{sql_query}{SQL query, defaulted to: \code{select * from table}.}

\item{parallel}{Leveraging the \code{\link{foreach}} package to export data by breaking it into smaller chunks.}

\item{n_core}{It is a required field if parallel is TRUE.}
}
\description{
Queries the data in an existing Domo DataSet.
}
\examples{
dataset_query(client_id = client_id,
  secret = secret,
  dataset_id = '3c907f1e-846c-4d06-88f6-35592db151f4',
  sql_query = 'select * from',
  parallel = TRUE,
  n_core = 7)
}
