% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRoDsUpdate.R
\name{pRoDsUpdate}
\alias{pRoDsUpdate}
\title{Create A New Dataset Via Stream API}
\usage{
pRoDsUpdate(
  client_id,
  secret,
  dataset_id,
  dataset,
  parallel = FALSE,
  n_core = NULL
)
}
\arguments{
\item{client_id}{A client_id that can be created on the developer.domo.com page.}

\item{secret}{A secret that can created on the developer.domo.com page.}

\item{nameLike}{If included, will limit the list of DataSets to those with names that contain the string passed in. nameLike is case insensitive.}

\item{sort}{The DataSet field to sort by. Fields prefixed with a negative sign reverses the sort (i.e. '-name' does a reverse sort by the name of the DataSets).}
}
\description{
It is best practice to use this function to create a new dataset as it leverages STREAM API that loads data parts at scale.
}
\examples{
dataset_get_info_all(client_id = client_id,
  secret = secret,
  nameLike = 'first',
  sort = 'name')
}
