% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRoPageGet.R
\name{pRoPageGet}
\alias{pRoPageGet}
\title{Retrieve All Pages}
\usage{
pRoPageGet(client_id, secret, data_frame = TRUE)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{data_frame}{If "TRUE", the output will be a data frame.}
}
\description{
Get a list of all pages in your Domo instance.
}
\examples{
pRoPageGet(client_id = client_id, secret = secret)
}
