% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRoPdp.R
\name{pRoPdp}
\alias{pRoPdp}
\title{Manage Personalized Data Permission (PDP) at Scale}
\usage{
pRoPdp(
  client_id,
  secret,
  data_table,
  parallel = FALSE,
  n_core = NULL,
  ignore_policy = "AA - Restricted"
)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{data_table}{A data frame/tibble that contains 5 required fields:
\enumerate{
 \item `Dataset ID`
 \item `Policy Name`
 \item `Policy Column`
 \item `User ID`
 \item `Policy Value`}
A combination of `Dataset ID`+`Policy Name`+`Policy Column` makes a unique key of the table and each unique key represents one policy.}

\item{parallel}{Leveraging the \code{\link{foreach}} package to simultaneously update the PDPs of multiple datasets.}

\item{n_core}{It is a required field if parallel is TRUE.}

\item{ignore_policy}{The operation will ignore all policies that contains the specified string, defaulted to 'AA - Restricted'.}
}
\description{
Manage Personalized Data Permission (PDP) at Scale with a master table that contains certain fields.
}
\examples{
data_table <- dplyr::tibble(`Dataset ID` = '824025cb-ba76-41b4-a38d-625a9cc1dfed',
  `Policy Name` = c('pol_1', 'pol_2'),
  `Policy Column` = c('col_1', 'col_2'),
  `User ID` = 123456789,
  `Policy Value` = c('val_1', 'val_2')))
pRoPdp(client_id = client_id,
  secret = secret,
  data_table = data_table)
}
