% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_create.R
\name{stream_create}
\alias{stream_create}
\title{Create a Stream}
\usage{
stream_create(client_id, secret, body)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{body}{The content of the dataset. Please refer to the documentation for details.}
}
\description{
When creating a Stream, specify the DataSet properties (name and description) and as a convenience, the create Stream API will create a DataSet for you.
In addition, you can only have one Stream open at a time. If you need to add additional data, we recommended adding more parts to the currently open Stream or executing a commit of the open stream before creating a new stream.
(Known limitation) The StreamAPI currently only allows you to import data to a DataSet created via the Stream API. For example, it is currently not supported to import data to a DataSet created by a Domo Connector.
\href{https://developer.domo.com/portal/7f71f4aa90726-create-a-stream}{DOMO Documentation}.
}
\examples{
columns <- list(); columns[[1]] <- list(name = 'Column 1', type = 'STRING')
schema <- list(columns = columns)
body <- list(dataSet = list(name = 'The Name', description = 'Some description', schema = schema), updateMethod = "REPLACE")
stream_create(client_id = client_id,
  secret = secret,
  body = body)
}
