% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_delete.R
\name{stream_delete}
\alias{stream_delete}
\title{Delete a Stream}
\usage{
stream_delete(client_id, secret, stream_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{stream_id}{The unique ID of the stream.}
}
\description{
\strong{This is destructive and cannot be reversed.} Deletes a Stream from your Domo instance. This does not a delete the associated DataSet.
\href{https://developer.domo.com/portal/6875057565875-delete-a-stream}{DOMO Documentation}.
}
\examples{
stream_delete(client_id = client_id,
  secret = secret,
  stream_id = 1)
}
