% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_execution_abort.R
\name{stream_execution_abort}
\alias{stream_execution_abort}
\title{Abort a Stream execution}
\usage{
stream_execution_abort(client_id, secret, stream_id, execution_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{stream_id}{The unique ID of the stream.}

\item{execution_id}{The unique ID of the execution.}
}
\description{
If needed during an execution, aborts an entire Stream execution. To abort the current stream execution within a Stream, simply identify the Stream’s ID within request
\href{https://developer.domo.com/portal/ab38f0e6a67ac-abort-a-stream-execution}{DOMO Documentation}.
}
\examples{
stream_execution_abort(client_id = client_id,
  secret = secret,
  stream_id = 1,
  execution_id = 1)
}
