% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_execution_commit.R
\name{stream_execution_commit}
\alias{stream_execution_commit}
\title{Commit a Stream execution}
\usage{
stream_execution_commit(client_id, secret, stream_id, execution_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{stream_id}{The unique ID of the stream.}

\item{execution_id}{The unique ID of the execution.}
}
\description{
Commits stream execution to import combined set of data parts that have been successfully uploaded.The Stream API only supports the ability to execute a “commit” every 15 minutes.
\href{https://developer.domo.com/portal/e326a44633f4c-commit-a-stream-execution}{DOMO Documentation}.
}
\examples{
stream_execution_commit(client_id = client_id,
  secret = secret,
  stream_id = 1,
  execition_id = 1)
}
