% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_execution_create.R
\name{stream_execution_create}
\alias{stream_execution_create}
\title{Create a Stream execution}
\usage{
stream_execution_create(client_id, secret, stream_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can created on the \url{developer.domo.com} page.}

\item{stream_id}{The unique ID of the stream.}
}
\description{
When you’re ready to upload data to your DataSet via a Stream, you first tell Domo that you’re ready to start sending data by creating an Execution.
Creating an Execution on a Stream will abort all other Executions on that Stream. Each Stream can only have one active Execution at a time.
\href{https://developer.domo.com/portal/242dd31d0746d-create-a-stream-execution}{DOMO Documentation}.
}
\examples{
stream_execution_create(client_id = client_id,
  secret = secret,
  stream_id = 1)
}
