% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_execution_get.R
\name{stream_execution_get}
\alias{stream_execution_get}
\title{Retrieve a Stream execution}
\usage{
stream_execution_get(client_id, secret, stream_id, execution_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{stream_id}{The unique ID of the stream.}

\item{execution_id}{The unique ID of the execution.}
}
\description{
To get the metadata of a specified stream execution.
\href{https://developer.domo.com/portal/e3d90b95387e7-retrieve-a-stream-execution}{DOMO Documentation}.
}
\examples{
stream_execution_get(client_id = client_id,
  secret = secret,
  stream_id = 1)
}
