% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_search.R
\name{stream_search}
\alias{stream_search}
\title{Search Streams}
\usage{
stream_search(client_id, secret, dataset_id, fields = "all")
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{fields}{Return desired fields: {all} or {id, dataset, updateMethod, createdAt, or modifiedAt}}

\item{qualifiers}{The search qualifiers to search by available qualifiers: dataSource.id or dataSource.owner.id}
}
\description{
Returns all Stream objects that meet argument criteria from original request.
\href{https://developer.domo.com/portal/41db49e09d006-search-streams}{DOMO Documentation}.
}
\examples{
stream_search(client_id = client_id,
  secret = secret,
  dataset_id = '3c907f1e-846c-4d06-88f6-35592db151f4',
  fields = 'all')
}
