% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_update.R
\name{stream_update}
\alias{stream_update}
\title{Update a Stream}
\usage{
stream_update(client_id, secret, stream_id, updateMethod)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{stream_id}{The unique ID of the stream.}

\item{updateMethod}{The data import behavior, either "REPLACE" or "APPEND".}
}
\description{
Updates the specified Stream’s metadata by providing values to parameters passed.
\href{https://developer.domo.com/portal/4930370cad24c-update-a-stream}{DOMO Documentation}.
}
\examples{
stream_update(client_id = client_id,
  secret = secret,
  stream_id = 1,
  updateMethod = "APPEND")
}
