% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_uploadPart.R
\name{stream_uploadPart}
\alias{stream_uploadPart}
\title{Upload a data part}
\usage{
stream_uploadPart(client_id, secret, execution_id, part_id, stream_id, body)
}
\arguments{
\item{client_id}{A client_id that can be created on the developer.domo.com page.}

\item{secret}{A secret that can created on the developer.domo.com page.}

\item{execution_id}{The ID of the Stream execution within the Stream.}

\item{part_id}{The ID of the data part being used to upload a subset of data within the Stream execution.}

\item{stream_id}{The ID of the Stream of data being imported into a DataSet.}

\item{body}{The content of the data. Please refer to the documentation for details.}
}
\description{
Creates a data part within the Stream execution to upload chunks of rows to the DataSet. The calling client should keep track of parts and order them accordingly in an increasing sequence. If a part upload fails, retry the upload as all parts must be present before committing the stream execution.
Parts can be uploaded simultaneously in separate threads assuming that each part has a distinct part ID and is ordered correctly. To reduce upload time, compress each data as a gzip file (application/gzip).
\href{https://developer.domo.com/portal/f186ddf30fca9-upload-a-data-part}{DOMO Documentation}.
}
\examples{
stream_uploadPart(client_id = client_id,
  secret = secret,
  execution_id = 1,
  part_id = 1,
  body = body)
}
