% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_create.R
\name{user_create}
\alias{user_create}
\title{Create A User}
\usage{
user_create(client_id, secret, body)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{body}{A list() of parameters that will be used to create the user, email, name, and role are required inputs aside from the following optional inputs: sendInvite, alternateEmail, employeeNumber, locale, location, phone, roleId, timezone, and title}
}
\description{
Creates a new user in your Domo instance.
\href{https://developer.domo.com/portal/6fc40f66c26e2-create-a-user}{DOMO Documentation}.
}
\examples{
user_create(client_id = client_id,
  secret = secret,
  body = list(name = 'John Doe', email = "email", role = "Admin", sendInvite = TRUE))
}
