% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_delete.R
\name{user_delete}
\alias{user_delete}
\title{Delete A User}
\usage{
user_delete(client_id, secret, user_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{user_id}{The unique ID of the user.}
}
\description{
\strong{This is destructive and cannot be reversed.} This function allows you to delete a certain user based on ID. Domo won't allow deletion of users if they own Datasets or Dataflows. You'll need to transfer ownership of these assets before deletion. All other assets previously owned by the deleted user will be orphaned.
\href{https://developer.domo.com/portal/b86a1b801b1f3-delete-a-user}{DOMO Documentation}.
}
\examples{
user_delete(client_id = client_id,
  secret = secret,
  user_id = '123456789')
}
