% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_get.R
\name{user_get}
\alias{user_get}
\title{Retrieve a user}
\usage{
user_get(client_id, secret, user_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{user_id}{The unique ID of the user}
}
\description{
Retrieves the details of an existing user.
\href{https://developer.domo.com/portal/7036c8cbcf865-retrieve-a-user}{DOMO Documentation}.
}
\examples{
user_get(client_id = client_id,
  secret = secret,
  user_id = '123456789')
}
