% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_get_all.R
\name{user_get_all}
\alias{user_get_all}
\title{List users}
\usage{
user_get_all(client_id, secret)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}
}
\description{
Get a list of all users in your Domo instance.
\href{https://developer.domo.com/portal/7f5648bdd673e-list-users}{DOMO Documentation}.
}
\examples{
user_get_all(client_id = client_id, secret = secret)
}
