% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_update.R
\name{user_update}
\alias{user_update}
\title{Update A User}
\usage{
user_update(client_id, secret, user_id, body)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{user_id}{The unique ID of the user.}

\item{body}{A list() of parameters that will be used to update the user, email is a required input aside from the following optional inputs: alternateEmail, employeeNumber, locale, location, name, phone, role, roleId, timezone, and title}
}
\description{
Updates the specified user by providing values to parameters passed. Any parameter left out of the request will cause the specific user’s attribute to remain unchanged. (Known limitations) Currently all user fields are required.
\href{https://developer.domo.com/portal/6d07be3b66fad-update-a-user}{DOMO Documentation}.
}
\examples{
user_update(client_id = client_id,
  secret = secret,
  user_id = '123456789',
  body = list(email = "email", role = "Admin"))
}
