% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appdb_doc_create.R
\name{appdb_doc_create}
\alias{appdb_doc_create}
\title{Create AppDB Document}
\usage{
appdb_doc_create(
  collection_id,
  developer_token,
  instance,
  data
)
}
\arguments{
\item{collection_id}{The AppDB collection ID.}

\item{developer_token}{A valid Domo Developer Token.}

\item{instance}{Domo instance URL. For example:
\code{"https://company.domo.com"}.}

\item{data}{A document payload, typically generated using
\code{appdb_row_to_document()}.}
}
\description{
Creates a new AppDB document using the provided payload.
}
\value{
The API response object returned by
\code{httr2::req_perform()}.
}
\examples{
new_data <- appdb_row_to_document(
  data = clients,
  row_number = 1
)

response <- appdb_doc_create(
  collection_id = "12345678-1234-1234-1234-123456789012",
  developer_token = developer_token,
  instance = "https://company.domo.com",
  data = new_data
)
}
