% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appdb_doc_delete.R
\name{appdb_doc_delete}
\alias{appdb_doc_delete}
\title{Delete AppDB Document}
\usage{
appdb_doc_delete(
  url,
  appdb_document_id,
  developer_token
)
}
\arguments{
\item{url}{The AppDB collection URL.}

\item{appdb_document_id}{The document ID to delete.}

\item{developer_token}{A valid Domo Developer Token.}
}
\description{
Deletes an existing AppDB document.
}
\value{
The API response object returned by
\code{httr2::req_perform()}.
}
\examples{
response <- appdb_doc_delete(
  url = collection_url,
  appdb_document_id = "12345678-1234-1234-1234-123456789012",
  developer_token = developer_token
)
}
