% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appdb_doc_update.R
\name{appdb_doc_update}
\alias{appdb_doc_update}
\title{Update AppDB Document}
\usage{
appdb_doc_update(
  url,
  appdb_document_id,
  developer_token,
  data
)
}
\arguments{
\item{url}{The AppDB collection URL.}

\item{appdb_document_id}{The document ID to update.}

\item{developer_token}{A valid Domo Developer Token.}

\item{data}{A document payload, typically generated using
\code{appdb_row_to_document()}.}
}
\description{
Updates an existing AppDB document.
}
\value{
The API response object returned by
\code{httr2::req_perform()}.
}
\examples{
update_data <- appdb_row_to_document(
  data = clients,
  row_number = 1
)

response <- appdb_doc_update(
  url = collection_url,
  appdb_document_id = "12345678-1234-1234-1234-123456789012",
  developer_token = developer_token,
  data = update_data
)
}
