% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appdb_row_to_document.R
\name{appdb_row_to_document}
\alias{appdb_row_to_document}
\title{Convert Data Frame Row to AppDB Document}
\usage{
appdb_row_to_document(
  data,
  row_number
)
}
\arguments{
\item{data}{A data frame containing the source data.}

\item{row_number}{The row number to convert.}
}
\description{
Converts a single row from a data frame into the
AppDB document structure required by the Domo AppDB API.
}
\value{
A list formatted for use with the AppDB API.
}
\examples{
new_data <- appdb_row_to_document(
  data = clients,
  row_number = 1
)

str(new_data)
}
`
