% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_get_all.R
\name{page_get_all}
\alias{page_get_all}
\title{List pages}
\usage{
page_get_all(client_id, secret, limit, offset)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{limit}{The amount of pages to return in the list. The default is 50 and the maximum is 500.}

\item{offset}{The offset of the page ID to begin list of pages within the response.}
}
\description{
Get a list of all pages in your Domo instance.
}
\examples{
page_get_all(client_id = client_id, secret = secret, limit = 500, offset = 0)
}
