% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_attachment_delete.R
\name{plt_attachment_delete}
\alias{plt_attachment_delete}
\title{Delete a project}
\usage{
plt_attachment_delete(client_id, secret, project_id, task_id, attachment_id)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can created on the \url{developer.domo.com} page.}

\item{project_id}{The unique ID of the project.}

\item{task_id}{The unique ID of the task.}

\item{attachment_id}{The unique ID of the attachment.}

\item{list_id}{The unique ID of the list.}
}
\description{
\strong{This is destructive and cannot be reversed.} Permanently deletes a project from your Domo instance.
\href{https://developer.domo.com/portal/abc4b606acabc-delete-an-attachment}{DOMO Documentation}.
}
\examples{
plt_attachment_delete(client_id = client_id,
  secret = secret,
  project_id = 123,
  list_id = 5,
  task_id = 12,
  attachment_id = 1)
}
