% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRoActivity.R
\name{pRoActivity}
\alias{pRoActivity}
\title{Retrieve Activity Log Entries (Simplified)}
\usage{
pRoActivity(client_id, secret, start = NULL, end = NULL, user_id = NULL)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{start}{(Optional) Only 'numeric' or 'Date' are acceptable as input. The start time(milliseconds) of when you want to receive log events, defaulted to the last 24 hours.}

\item{end}{(Optional) Only 'numeric' or 'Date' are acceptable as input. The end time(milliseconds) of when you want to receive log events, defaulted to now.}

\item{user_id}{(Optional) Specify the ID of the user to retrieve a certain individual's activities.}
}
\description{
This function allows you to retrieves all activity log entries, all time/date values are in UTC.
}
\examples{
pRoActivity(client_id = client_id,
  secret = secret,
  start = as.Date('2024-01-01'),
  end = as.Date('2024-01-01'),
  user_id = 27)
}
