% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRoDsUpdate.R
\name{pRoDsUpdate}
\alias{pRoDsUpdate}
\title{Update A Dataset Via Stream API}
\usage{
pRoDsUpdate(
  client_id,
  secret,
  dataset_id,
  dataset,
  parallel = FALSE,
  n_core = NULL
)
}
\arguments{
\item{client_id}{A client_id that can be created on the \url{developer.domo.com} page.}

\item{secret}{A secret that can be created on the \url{developer.domo.com} page.}

\item{dataset_id}{The unique ID of the dataset.}

\item{dataset}{The data frame that will be uploaded to DOMO.}

\item{parallel}{(Optional) If TRUE, the function will use multiple cores to upload the data frame. `n_core` is required when `parallel` = TRUE.}

\item{n_core}{(Optional) The number of cores to use from your device. It's recommended to leave one core free. If you have 8 cores, enter `n_core` = 7.}
}
\description{
It is best practice to use this function to update an existing API dataset as it leverages STREAM API that loads data parts at scale.
}
\examples{
pRoDsUpdate(client_id = client_id,
  secret = secret,
  dataset_id = '3c907f1e-846c-4d06-88f6-35592db151f4',
  dataset = dataset)
}
